requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
base58==2.1.1
bitarray==2.7.3
black==23.1.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cytoolz==0.12.1
Django==4.1.7
eth-abi==2.2.0
eth-account==0.5.9
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b7
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
futureproof==0.3.1
hexbytes==0.2.3
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.4
netaddr==0.8.0
packaging==23.0
parsimonious==0.8.1
pathspec==0.11.0
platformdirs==3.1.1
protobuf==3.19.5
psutil==5.9.4
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.11.0
py-solc==3.2.0
py-solc-x==1.1.1
pycodestyle==2.10.0
pycryptodome==3.17
pyethash==0.1.27
pyrsistent==0.19.3
pysha3==1.0.2
python-dotenv==1.0.0
regex==2022.10.31
requests==2.28.2
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
toolz==0.12.0
trie==2.0.0a5
typing-extensions==3.10.0.2
ujson==5.7.0
urllib3==1.26.15
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
